chore: move postgres to separate module
