+++ dataplane-sdk-postgresql/src/test/java/org/eclipse/dataplane/store/postgresql/PostgresControlPlaneStoreTest.java
@@ -12,11 +12,10 @@
  *
  */
 
-package org.eclipse.dataplane.store.sql;
+package org.eclipse.dataplane.store.postgresql;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.eclipse.dataplane.port.store.ControlPlaneStore;
-import org.eclipse.dataplane.port.store.sql.PostgresControlPlaneStore;
 import org.eclipse.dataplane.store.ControlPlaneStoreTestBase;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
