+++ settings.gradle.kts
@@ -8,5 +8,6 @@ dependencyResolutionManagement {
 
 include(":dataplane-sdk-core")
 include(":dataplane-sdk-jakarta-ee")
+include(":dataplane-sdk-postgresql")
 
 include(":e2e-tests")
